chore: refactor interactive templating to be importable
This makes it easier for interactive-templates to import the code to use
in tests
